+++ src/flockwave/server/config.py
@@ -65,7 +65,7 @@ EXTENSIONS = {
                 "connections": [
                     "default"
                 ],  # default setup; listens for heartbeats on UDP port 14550, sends broadcasts to UDP port 14555
-                "routing": {"rtk": 0},
+                "routing": {},
             }
         },
     },